backend/pyproject.toml
[project]
name = "docflow-backend"
version = "0.1.0"
description = "Plateforme de traitement automatique de documents administratifs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pymongo>=4.15.0",
    "pypdf>=6.8.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "uvicorn[standard]>=0.41.0",
    "motor>=3.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.0.0,<4.0.0",
    "email-validator>=2.1.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
